mv service code to its own pkg
