+++ cmd/servicebindingcleanup/main.go
@@ -2,44 +2,29 @@ package main
 
 import (
 	"context"
-	"fmt"
 	"log/slog"
 	"os"
 	"time"
 
-	"github.com/kyma-project/kyma-environment-broker/internal"
 	"github.com/kyma-project/kyma-environment-broker/internal/broker"
 	"github.com/kyma-project/kyma-environment-broker/internal/events"
 	"github.com/kyma-project/kyma-environment-broker/internal/schemamigrator/cleaner"
+	"github.com/kyma-project/kyma-environment-broker/internal/servicebindingcleanup"
 	"github.com/kyma-project/kyma-environment-broker/internal/storage"
 	log "github.com/sirupsen/logrus"
 	"github.com/vrischmann/envconfig"
 )
 
-type BrokerClient interface {
-	Unbind(binding internal.Binding) error
+type Config struct {
+	Database storage.Config
+	Broker   broker.ClientConfig
+	Job      JobConfig
 }
 
-type Config struct {
-	Database       storage.Config
-	Broker         broker.ClientConfig
+type JobConfig struct {
 	DryRun         bool          `envconfig:"default=true"`
 	RequestTimeout time.Duration `envconfig:"default=2s"`
 	RequestRetries int           `envconfig:"default=2"`
-}
-
-type ServiceBindingCleanupService struct {
-	cfg             Config
-	brokerClient    BrokerClient
-	bindingsStorage storage.Bindings
-}
-
-func newServiceBindingCleanupService(cfg Config, client BrokerClient, bindingsStorage storage.Bindings) *ServiceBindingCleanupService {
-	return &ServiceBindingCleanupService{
-		cfg:             cfg,
-		brokerClient:    client,
-		bindingsStorage: bindingsStorage,
-	}
 }
 
 func main() {
@@ -51,19 +36,19 @@ func main() {
 	var cfg Config
 	fatalOnError(envconfig.InitWithPrefix(&cfg, "APP"))
 
-	if cfg.DryRun {
+	if cfg.Job.DryRun {
 		slog.Info("Dry run only - no changes")
 	}
 
 	ctx := context.Background()
-	brokerClient := broker.NewClientWithRequestTimeoutAndRetries(ctx, cfg.Broker, cfg.RequestTimeout, cfg.RequestRetries)
+	brokerClient := broker.NewClientWithRequestTimeoutAndRetries(ctx, cfg.Broker, cfg.Job.RequestTimeout, cfg.Job.RequestRetries)
 	brokerClient.UserAgent = broker.ServiceBindingCleanupJobName
 
 	cipher := storage.NewEncrypter(cfg.Database.SecretKey)
 	db, conn, err := storage.NewFromConfig(cfg.Database, events.Config{}, cipher, log.WithField("service", "storage"))
 	fatalOnError(err)
 
-	svc := newServiceBindingCleanupService(cfg, brokerClient, db.Bindings())
+	svc := servicebindingcleanup.NewService(cfg.Job.DryRun, brokerClient, db.Bindings())
 	fatalOnError(svc.PerformCleanup())
 
 	slog.Info("Service Binding cleanup job finished successfully!")
@@ -71,29 +56,6 @@ func main() {
 	fatalOnError(conn.Close())
 	logOnError(cleaner.HaltIstioSidecar())
 	fatalOnError(cleaner.Halt())
-}
-
-func (s *ServiceBindingCleanupService) PerformCleanup() error {
-	slog.Info(fmt.Sprintf("Fetching Service Bindings with expires_at <= %q", time.Now().UTC().Truncate(time.Second).String()))
-	bindings, err := s.bindingsStorage.ListExpired()
-	if err != nil {
-		return err
-	}
-
-	slog.Info(fmt.Sprintf("Expired Service Bindings: %d", len(bindings)))
-	if s.cfg.DryRun {
-		return nil
-	} else {
-		slog.Info("Requesting Service Bindings removal...")
-		for _, binding := range bindings {
-			err := s.brokerClient.Unbind(binding)
-			if err != nil {
-				slog.Error(fmt.Sprintf("while sending unbind request for service binding ID %q: %s", binding.ID, err))
-				continue
-			}
-		}
-	}
-	return nil
 }
 
 func fatalOnError(err error) {
